off_t and seek related fixes

+++ src/misc/sockmod.c
@@ -112,7 +112,7 @@ static void read_file(CTX, struct fbuf* fb, const char* name)
 
 	if((rd = sys_read(fd, buf, st.size)) < 0)
 		fail("read", name, rd);
-	if((size_t)rd < st.size)
+	if(rd < st.size)
 		fail("incomplete read from", name, 0);
 
 	buf[st.size] = '\0';
